requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==4.2.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
bitsandbytes==0.42.0
blinker==1.9.0
blis==1.2.0
Brotli==1.1.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cloudpathlib==0.20.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.5.14
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
duckduckgo_search==7.4.2
durationpy==0.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
entrypoints==0.4
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.7.4
fastapi==0.115.8
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.1.24
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
gliner==0.2.17
google-auth==2.38.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
kubernetes==32.0.0
langchain==0.3.18
langchain-chroma==0.2.1
langchain-community==0.3.17
langchain-core==0.3.34
langchain-experimental==0.3.4
langchain-huggingface==0.1.2
langchain-neo4j==0.4.0
langchain-ollama==0.2.3
langchain-openai==0.3.5
langchain-text-splitters==0.3.6
langcodes==3.5.0
langsmith==0.3.6
language_data==1.3.0
libmagic==1.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
neo4j==5.28.1
neo4j-graphrag==1.6.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
# onnxruntime==1.20.1
openai==1.61.0
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
posthog==3.11.0
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
Pympler==1.1
pyparsing==3.2.1
pypdf==5.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentence-transformers==3.4.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.5.1
stack-data==0.6.3
starlette==0.45.3
streamlit==1.22.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tldextract==5.1.3
tokenizers==0.21.0
toml==0.10.2
toolz==1.0.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
typer==0.15.1
types-PyYAML==6.0.12.20250402
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
unstructured==0.9.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
validators==0.34.0
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==14.2
widgetsnbextension==4.0.13
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
yfiles_jupyter_graphs==1.10.1
zipp==3.21.0
zstandard==0.23.0
